Add service.PermissionsStore interface
The PermissionsStore interface was defined only in the API package, which meant it was restricted to what the API package required. The PermissionsStore is also used in the permissions package, outside of the API. Having the service define its own Store interface prevents it from being tied to the API package.